+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - add `item-service.serializeItem(item)` internal function that will correctly serialize an item for POSTing to the item apis via jQuery
 - add `portal-service` with `update`
 - bumped to node 6.3.0 in `.nvmrc` and `.travis.yml`
+- fix bugs/typos in sharing-service 
 
 ## [0.2.4]
 ### Added